Adds cli -F (flatten all) and -f (flatten k1,k2,...)
These changes rely heavily on the cli developed in jsoncut.
Considering writing code that does not rely on jsoncut.cli.
This commit does not add "-g" functionality just yet on the cli,
which will generate rows for each item in a Sequence, so that it
can be saved in a CSV file, or similar.