Refines the cli interface and handles arrays of json as input.
Gets rid of the -F (--Flatten) option and instead flattens the entire
object when no args are set for the -f option.  -f is now allowed multiple
arguments and can be called like: `-f7,9` or `-f7 -f9`.  In either case
both keys 7 and 9 are flattened.  Lastly, if the input is an array of JSON,
then the cli handles each item in the array as a separate JSON document and
flattens it appropriately for each item in the array.